pyproject.toml
[project]
name = "ai-doppelganger"
version = "0.1.0"
description = "Single-user AI doppelganger API"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0,<1.0.0",
    "google-api-python-client>=2.0.0,<3.0.0",
    "google-auth-httplib2>=0.2.0,<1.0.0",
    "google-auth-oauthlib>=1.2.0,<2.0.0",
    "httpx>=0.27.0,<1.0.0",
    "openai-agents",
    "pydantic>=2.0.0,<3.0.0",
    "python-dotenv>=1.0.1,<2.0.0",
    "uvicorn[standard]>=0.30.0,<1.0.0",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project.optional-dependencies]
test = [
    "pytest>=8.0.0,<9.0.0",
    "pytest-asyncio>=0.23.0,<1.0.0",
]

[tool.setuptools.packages.find]
include = ["app*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
